Added tests for examples 1, 2, 3 and 4
The code of the testing was refactorized to easily expand it to
test other examples. A module `examples.py` contain separate functions
to generate the input for each example. The fixture `example_data`
is parametrized to receive the function to run.

Most of the tests are run with a combinatorial product of all
example cases and all speeds, and the result is tested against
expected results pre-generated for each case and stored in json

All those json were moved to tests/expected_sols and the names
of the files is structured for easier identification and
programatically access.